Login en BD

+++ proyecto/app/services/GenericCrud.jsx
@@ -1,19 +1,16 @@
 import oracledb from 'oracledb';
 import { objectToString } from '../functions/objectToString';
-
+const ConnectDBUtil = require('./Coneccion.jsx');
 class GenericCrud {
     constructor(tableName) {
         this.tableName = tableName;
+        this.connectDB = ConnectDBUtil.connectDB;
     }
 
     async openConnection() {
         try {
-            return await oracledb.getConnection({
-                user: process.env.DB_USER,
-                password: process.env.DB_PASSWORD,
-                connectionString: process.env.DB_CONNECTION_STRING,
-                privilege: oracledb[process.env.DB_ROLE]
-            });
+            console.log("33333333333333333333333", ConnectDBUtil.connectDB);
+            return await oracledb.getConnection(ConnectDBUtil.connectDB);
         } catch (err) {
             console.error("Error de conexión:", err);
             throw err;
@@ -23,12 +20,10 @@ class GenericCrud {
     async create(data) {
         const connection = await this.openConnection();
         try {
-            console.log(data);
             const columns = Object.keys(data).join(', ');
             const values = Object.values(data).map(value => `${value != "NULL" ? `'${value}'` : `${value}` }`).join(', ');
 
-            const sql = `INSERT INTO ${this.tableName} (${columns}) VALUES (${values})`;
-            console.log(sql);
+            const sql = `INSERT INTO admin.${this.tableName} (${columns}) VALUES (${values})`;
             const result = await connection.execute(sql);
             await connection.commit();
 
@@ -41,7 +36,7 @@ class GenericCrud {
     async read(whereClause = '') {
         const connection = await this.openConnection();
         try {
-            const sql = `SELECT * FROM ${this.tableName} ${whereClause}`;
+            const sql = `SELECT * FROM admin.${this.tableName} ${whereClause}`;
             const result = await connection.execute(sql);
     
             const columns = result.metaData.map(column => column.name);
@@ -52,6 +47,7 @@ class GenericCrud {
                 });
                 return rowData;
             });
+            
     
             return rows;
         } finally {
@@ -65,7 +61,7 @@ class GenericCrud {
         try {
             const updates = objectToString(data);
 
-            const sql = `UPDATE ${this.tableName} SET ${updates} WHERE ${whereClause}`;
+            const sql = `UPDATE admin.${this.tableName} SET ${updates} WHERE ${whereClause}`;
             const result = await connection.execute(sql);
             await connection.commit();
 
@@ -78,7 +74,7 @@ class GenericCrud {
     async delete(whereClause) {
         const connection = await this.openConnection();
         try {
-            const sql = `DELETE FROM ${this.tableName} WHERE ${whereClause}`;
+            const sql = `DELETE FROM admin.${this.tableName} WHERE ${whereClause}`;
             const result = await connection.execute(sql);
             await connection.commit();
             return result;
@@ -86,6 +82,56 @@ class GenericCrud {
             await connection.close();
         }
     }
+
+    async chageConection(codigo, password) {
+        console.log(codigo, password);
+        console.log("11111111111111111111111111", ConnectDBUtil.connectDB);
+        const cod = "" + codigo;
+        const pas = ""+password;
+        console.log("11111111111111111111111111", cod, " ", pas);
+        ConnectDBUtil.updateConnectDB({
+            user: codigo,
+            password: password,
+            connectionString: process.env.DB_CONNECTION_STRING,
+            privilege: oracledb[""]
+          });
+        console.log("2222222222222222222222", ConnectDBUtil.connectDB);
+    }
+
+    async userRole(whereClause = '') {
+        const connection = await this.openConnection();
+        try {
+            const sql = `SELECT GRANTED_ROLE FROM DBA_ROLE_PRIVS ${whereClause}`;
+            const result = await connection.execute(sql);
+    
+            const columns = result.metaData.map(column => column.name);
+            const rows = result.rows.map(row => {
+                const rowData = {};
+                row.forEach((value, index) => {
+                    rowData[columns[index]] = value;
+                });
+                return rowData;
+            });
+            
+    
+            return rows;
+        } finally {
+            await connection.close();
+        }
+    }
+    async logout(){
+        console.log("Putaaaaaaaaaaaaaaaaaaaaaamadreeeeeeeeeee");
+        ConnectDBUtil.updateConnectDB(
+            {
+                user: process.env.DB_USER,
+                password: process.env.DB_PASSWORD,
+                connectionString: process.env.DB_CONNECTION_STRING,
+                privilege: oracledb[process.env.DB_ROLE]
+              }
+        );
+       
+    }
+      
 }
 
 export default GenericCrud;
